Added a reusable NSNumber boolean value transformer
As collections can only contain objects, booleans are represented by
NSNumber. A number will be used when serializing to JSON, despite the
latter's support for booleans, except when the NSNumber is backed by
NSCFBoolean, where true booleans are used.

This transformer ensures that the NSNumber is backed by an NSCFBoolean.