Utilize the power of toll-free bridges.
Refactoring MTLBooleanValueTransformerName to return instances of
CFBooleanRef; specifically the only two: kCFBooleanTrue and
kCFBooleanFalse, which are toll-free bridged to NSNumber.

Added a comment to explain why one would want to use
MTLBooleanValueTransformerName. In addition to NSJSONSerialization,
I believe it true that NSUserDefaults suffer from this same issue:
though I'll love to be proved false.

The previous tests were, er, fragile. I blame the fact that said commit
happened at nearly four-thirty in the morning. Instead of testing
against the className (or the description of the class, I couldn't
decide which…), test against the corresponding CFBooleanRef values.