Add a cache for transformer objects.
In my tests this greatly speeds up serialization/deserialization. The transformer objects are cached instead of being created/deleted all the time. We also save time by not requiring the NSInvocation and building up the selector. This is thread safe as well since we atomically save the cache object; in the worst case we would look up transformers more than once.

Downside: The cache is saved on the class object, so it won't ever be cleaned up. I don't see this as an issue though as transformers usually are very small objects.