Store managed object relationships as arrays
Since models in Core Data are uniqued by the ObjectID. You can have
multiple object's with the same contents, but different managed object IDs.
However -[MTLModel isEqual:] method will say that two objects, with the same
contents are the same (because they are in this context). When you
convert these different managed object's into MTLModels and then place
them in an NSSet, it will treat them as equal and only show one item.

Closes #352