Ensure every model key is validated when instantiating from a dictionary
Currently only the keys actually present in the dictionary are validated
during initialization. This means that if a key is not present in the
dictionary, a non-nil property will still be initialized to a nil value.

Expanded the set of keys validated during initialization to include all
the class's property keys, in addition to the keys found in the
dictionary.