Update JSONDictionaryFromModel: method in MTLJSONAdapter This update addresses a deficiency in Mantle 2.0 where work done in a model subclass dictionaryValue: method is ignored during serialization.
A side effect of this update is that any keys that are _added_ via serializablePropertyKeys: will be ignored. However any keys that are removed via this method will be removed during serialization. The assumption is that adding a key in serializablePropertyKeys: isn't nearly as useful because the model would have no way to map a value to that new key (unless work in dictionaryValue is done as well).

The biggest thing this update solves is backwards compatibility as it respects any work previously being done in dictionaryValue:. Also this removes the requirement to subclass the adapter anytime serialization manipulation needs to be done (like dropping null values). This is particularly useful because the models are already subclasses and come with a purpose built method for manipulating their dictionary representation. And thus can be done on a per model basis or handled in a base model class that overrides dictionaryValue and handles default manipulation logic.