Check superclass for property persistence
A subclass can shadow a parent's property declaration, either explicitly
with a `@property` directive or implicitly by conforming to a protocol
that requires a particular property. The result is that the subclass
property record has no associated ivar (or methods), and calls to get /
set the property will fall up to the superclass's methods (backed by the
superclass's ivar).